recipe-generator/requirements.txt
PyYAML>=6.0
pytest>=7.0
pykickstart>=1.99
